Bug#14784707 FAILED TO CONVERT TRANSPORTER ERRORS UNDER CONNECTION LOAD
Bug report complains about WARNINGS seen in the MgmtSrvr log:

2022-12-05 09:55:04 [MgmtSrvr] WARNING  -- Failed to convert connection from '100.103.27.109:38400' to transporter:
 Ignored connection attempt as this node is not expecting a connection from node 2. State 2

The expectation seems to be that a WARNING should not be
logged unless there is something abnormal going on.
Bug report also claims that 'events' are lost around the
time of mgmd restarts. Which is likely a different matter,
but the WARNING might have drawn the attention in the wrong
direction.

The warning messages is produced from within TransporterRegistry::connect_server()
when it finds that the transporter it tries to connect to has
performStates[nodeId] == DISCONNECTED. This is an expected state right after
a node has been restarted, where the asynchronous setup of the transporter connection
has not yet completed all the required steps. This is one of several
reasons that connect_server() may fail, some more serious than the others.

There is a previous patch which attempts to classify how serious the
connect reject reasons are (A side effect of other issues fixed in patch):

commit cb105dd29b08350ac633e366e7f3a6918646e7e7
Merge: 471490d69a3 78707571cf1
Author: Frazer Clement <frazer.clement@oracle.com>
Date:   Thu Oct 24 16:21:38 2019 +0100

    Bug #30368622 NODES ARE CRASHING DURING RESTART WITH ERROR ""ANOTHER NODE FAILED ...

That patch also introduced a 'bool& log_failure' argument to ::connect_server()

bool
 TransporterRegistry::connect_server(NDB_SOCKET_TYPE sockfd,
                                     BaseString & msg,
-                                    bool& close_with_reset)
+                                    bool& close_with_reset,
+                                    bool& log_failure)
 {
   DBUG_ENTER("TransporterRegistry::connect_server(sockfd)");

For connection attempts with 'performStates[nodeId] == DISCONNECTED'
log_failure will be returned as 'false':

+      /**
+       * This is expected if the transporter state is DISCONNECTED,
+       * otherwise it's a bit strange
+       */
+      log_failure = (performStates[nodeId] != TransporterRegistry::DISCONNECTED);

The management server didn't take advantage if this now argument:

@@ -4905,7 +4905,8 @@ MgmtSrvr::transporter_connect(NDB_SOCKET_TYPE sockfd,
 {
   DBUG_ENTER("MgmtSrvr::transporter_connect");
   TransporterRegistry* tr= theFacade->get_registry();
-  if (!tr->connect_server(sockfd, msg, close_with_reset))
+  bool dummy_log_failure = false;
+  if (!tr->connect_server(sockfd, msg, close_with_reset, dummy_log_failure))
     DBUG_RETURN(false);

This patch lets MgmtSrvr::transporter_connect() now return 'log_failure' upto
the callee MgmApiSession::transporter_connect(), which is where the warning
is written to the MgmdSrvr logs. The returned warning message will now only
be written if 'log_failure' is set.

This is similar to how ::connect_server() failures are (not) written to
the data node logs when TransporterService::newSession() attempt to
::connect_server().

Change-Id: Ibc83403a8a01a95998357b29f51b3150accd7b7d