Bug#34854369: Customer query hits assert(m_pq.is_valid()) failure [3/3, MyISAM]
An assert failure was seen in the hypergraph optimizer during graph
simplification when running some queries with joins that involved
empty MyISAM tables.

Like in the queries fixed by the previous two patches, the assertion
was caused by row estimates that were zero, which led to division by
zero and meaningless results.

Whereas InnoDB is more cautious and never estimates the cardinality of
a table to zero, MyISAM does return an estimate of zero if the table
is empty. And the graph simplification code is not prepared for this.

Fixed by making graph simplification adjust row estimates of zero to
one before doing calculations with them. Even checking if a table is
empty has a cost, so assuming at least one read from each table sounds
reasonable.

Change-Id: I9f043a47435d87cb5756d8abf1f7e50c2ef61c30