Enh 34110068 - Request to get more timestamps into the output of ndb_restore
Prefix the ndb_restore outputs with a timestamp, managed by the new bool option :
"--timestamp_printouts".

By default, it is 'false' in in 7.5/7.6.

So there is no changes to tests. However, one test
(ndb_restore_print_sql_log) is changed to use the option to 'true' in
order to test that it is working. The test is also changed to mask
timestamps from the restore output with 'X' to avoid tests running on
different dates give 'Result content mismatch'.

Reviewed by: Martin Skold <Martin.Skold@oracle.com>

Change-Id: Ic183206e4d0f1d065d6448d03172a04217d4f4a0