Bug#34877154 routertest_component_rest_api_enable uses fixed port nums (random failures)
The test sometimes runs a bootstrap and launches the Router with
produced config file. The problem is that the default Router accepting
ports (6446 etc.) are sometimes used, which can collide with other
tests or processes in the OS.

This patch makes sure this never happens. Always overwrites the defaults
if the configuration is going to be used for launching the Router.

Change-Id: I94db246712ec97bfed94037eecacc8aff4efc10d