Bug#34145006: SE private ids of different SEs may be mixed when marked as not found
Problem: DD has code which caches failed lookups of se_private_ids (ids
which are not found). This was done to speed up Innodb specific code and
relies on the fact that InnoDB will not reuse its se_private_ids. This
assumption does not necessarily hold for other storage engines. Notably
it does not for NDB where this problem was fixed in bug#33824058, by not
using this cache.

Solution: The idea from the NDB fix is extended so that the cache lookup
is only used when the engine is InnoDB. This is a simpler solution than
trying to make se_private_ids from different engines co-exist in the
same cache.

Change-Id: I1b5e6cad12325acb131eabd79a5b93967315edf5