BUG#34849343 Aligned_atomic not working as intended - aligned_alloc
Problem
-------
The Aligned_atomic class in sql/memory/aligned_atomic.h was created with the
intention to align and pad a std::atomic<T> object so that it starts at the
beginning of a cache line and is padded so that it fills the entire cache
line. The intended use case was to ensure that an array of Aligned_atomic
objects keeps the elements in contiguous memory but different cache lines,
eliminating false sharing.

Aligned_atomic is implemented by constructing a new char* buffer the size of
a cache line, and then in-place-construct a std::atomic<T> at the beginning
of the buffer. It keeps a std::atomic<T>* as a member; this member is aligned
to the cache line size.

This only has performance impact, no functional defect.

Analysis / Root-cause analysis
------------------------------
This implementation does not seem meet the intentions, AFAIU:
- There is a call to `new` for each Aligned_atomic object. So the elements
are not contiguous.
- The pointer to the buffer is aligned, but not the buffer itself. So it is
possible that the buffer begins in the middle of a cache line and extends
into the next cache line. In fact, if the alignment is smaller than the size
of the atomic type, an std::atomic may intersect with two cache lines, and if
two such objects end up adjacent to each other, there can be false sharing.
More likely, there can be false sharing between the Aligned_atomic and some
object of another type.

The major drawback is probably that the construction of an array of
Aligned_atomic will allocate many small objects instead of one big, which
uses more memory and CPU than necessary.

Solution
--------
Object std::atomic<T> is allocated using placement new operator in aligned memory.

Change-Id: I3a5f002959137a8d2a47b1dd6229b2d5cb82a030