WL#15166 patch #4 Win32 fixes
On Win32, every binary needs one instance of openssl/applink.c.

MySQL has one in client_authentication.cc, and this one is present
in the server and in the client library.

This patch adds instances to ndb_sign_keys, ndb_mgmd, ndbd,
and testNodeCertificate-t, and includes fixes for other
assorted compiler errors and warnings on Win32.

Change-Id: I2d7f940b92ddac7d860d2c6fc2d98ead23e195b2