Bug#34317267: Assertion `!table->file->inited' failed in Mysql 8.0.29.
The problem here is that we attempt to perform an optimization of
a MAX/MIN aggregation on an outer reference when optimizing a
join condition for the outer query block, and the table this field
belongs to has been const-optimized. In this case, the table has not
yet been closed as expected by the max/min aggregation code, and this
assert is hit.

The fix is to prevent this optimization for fields that are outer
references and that have been or are in the process of being const-
optimized. It might be possible to handle such fields better, however
the added complexity does not seen worthwhile. In any case, outer
references that are not const-optimized are also not max/min optimized.

Change-Id: Ifd99e104171ba11f74c77be1376608cefc281c91