Bug#36159678 (1 of 6) JTie: remove some larger macros in jtie_gcalls
In jtie_gcalls.hpp, macros are used to create lists and templates.
Many of these could stringify or concatenate up to 20 parameters,
but capacity beyond about 13 is not needed.

This patch shortens the lists. Clang will sometimes print a
"rejecting template" error for each candidate template not chosen,
so this helps to reduce the number of error messages generated
before the compiler finally gives up.

Change-Id: Iddbb25d0a6e5740815b686259c734a0bbf434d8d