Bug#36159678 (2 of 6) JTie: Restore const variants of gcall templates
An earlier version of jtie_gcalls generated two distinct templates
for gcall_mfv() and gcall_mfr() to map the const and non-const
variants of otherwise identical method calls.

This strategy was abandoned because some very old C++ compilers could
not handle the macro invocations that were used, and because
Microsoft's compilers considered the two templates to be equivalent.

Microsoft compilers still behave the same way, but Clang requires
distinct templates. This patch restores the earlier design, but
only in non-Microsoft environments.

Change-Id: I6694c7cdc096a6675e71bcd843fec99421367c70