Bug#34099408: Result mismatch seen with DENSE_RANK window               function
When resolving rollup, expressions are wrapped with rollup
functions if needed. For a window function, since they are
processed after rollup, all arguments to it should be
updated once rollup wrappers are added. In case of rank
functions, resolver updates its order by expressions.
However, order by expressions of a rank function are
wrapped in Item_ref objects (used in ref slices) and
as a result, when Item_ref gets replaced with the
changed order by expression, it gives wrong results.

The function which updates the arguments to window functions
is called only when resolving rollup for old optimizer.
However for hypergraph optimizer it gets called from several
different places and it expects that the Item_ref be replaced
with the changed order by expression. So the solution is to
replace the item referenced by the Item_ref object in
case of the old optimizer and the reference itself in case
of hypergraph.

Change-Id: I733c8bab1d72e69f94ac6790bf6c55d3b8b352fb