Bug#29642686 CUSTER.SETOPTION("CLUSTERNAME","MYNEWCLUSTERNAME") BREAKS INNODB CLUSTER
While bootstrapping the Router it writes the cluster name to the static
configuration file. It is later used in SQL 'where' clauses when
querying the metadata, to make sure the same Cluster is used than
the one used for bootstrap. This results in errors once the users
change the name of the Cluster (which Shell allows). The UUID should
be used for that purpose instead of a cluster name (GR UUID in case
of GR cluster, ClusterSet UUID in case of ClusterSet and Cluster UUID
in case of ReplicaSet Cluster).

This patch changes the logic so that the UUID written to the state
json file is used to query the metadata. The cluster name is only used
when the UUID is not in the state file (back in the days it was not
stored there, so very old configurations will not have it).
The cluster_name option is kept in the config file for better backward
compatibility.

Change-Id: I4b87b1f2e4394296c75cba36f61a027561e6f7c6