WL#12794 rw-splitting [07/17] - lazy-connect, check super-read-only
read/write splitting needs to ensure that the read-only servers are
really "super-read-only" to ensure that a statement that's sent to
read-only server does not break the consistency of the cluster
if the server happens to be writable.

This could happen if the cluster is using Async Replication only
without any of the safe-guards that Group Replication provides.

A read-write statement may be targetted at a read-only server by
setting the access_mode explicitly:

  ROUTER SET access_mode='read_only';
  INSERT INTO tbl VALUES (...);

Change
------

Check that the read-only server has @@super_read_only set.

Fallback to read-write server if not.

Change-Id: I7e31295063fd4a5afc5160d739fa2afd9f1b3f47