Bug#34594825 NdbSpin_Init calculates wrong number of loops in NdbSpin Bug#32497174 NDBSPIN_INIT LOGIC PROBLEMS
NdbSpin_Init() is intended to find how many times we need to call
cpu_pause() in order to wait for 1 microsecond. It will set
this loop-count into the variable glob_num_spin_loops, which
is later used by NdbSpin() which will 'pause' this number
of times to do the ~1us wait.

Initially we set 'glob_num_spin_loops = 10' as a guess,
which would also cause NdbSpin() to 'pause' 10 times
until we have calibrated the spin-loop with NdbSpin_Init()

However, NdbSpin_Init() did its calibration loop by calling
NdbSpin(), such that what was measured is the time it takes
to cpu_pause 10 times, not the cpu_pause() itself.

Patch fix this by replacing the NdbSpin() call in 'init'
with cpu_pause().

It also fix issue 2) described in bug#32497174, by introducing
the new variable 'glob_spin_enabled = false'. This variable is
set to 'true' only after a successfull NdbSpin_Init(), finding
a non-zero elapsed time for the nanos_per_call.
NdbSpin_is_supported() is also changed to return this variable,
requiring a successfull NdbSpin_Init() before Spin-support is
claimed.

Furthermore some adjustments were done to places where NdbSpin()
was called, using the assumption that the code was adapted to
the too short 100ns spin.

  - Places where we had to NdbSpin() directly after each other
    were changed to a singel NdbSpin()
  - Places where we added twe more 'num_spins' to be looped over
    for the next wait were changed to a '++' instead.

Also removed code from mt.cpp where we claim to only check
the elapsed time for every 3 times we NdbSpin(). As the spin is
now 10x longer it should be ok to check elapsed every time.

Note that the removed code was incorrect and didn't do
what it claimed to do:

-      /* Check if we have done enough spinning once per 3 us */
-      if ((i & 3) == 3)
-        continue;

... In reality we checked 3 of 4 times, instead of skipping
the test 3 of 4 times.

Change-Id: I77e5577025ee2c96dc10d4ae4bada3006121655a