Bug#34870181: GR operations hang by accessing replication_group_communication_information, replication_group_member_stats in parallel
Description:
------------
STOP GROUP_REPLICATION hangs when
replication_group_communication_information, and
replication_group_member_stats performance schema tables are accessed
in parallel. This causes SELECT against these performance schema tables
hang, too.

Analysis and Fix:
-------------------
The hang happens when STOP GROUP_REPLICATION and
replication_group_communication_information are queried in paralel and
STOP GROUP_REPLICATION is able to run before the query and the XCom
thread is stopped as result of the operation.

When the replication_group_communication_information query tries to
request the value for the WRITE_CONCURRENCY, we are able to add a
request to the queue between GCS and XCom, but the code was not prepared
to not to have an answer.

We modified Gcs_xcom_group_management::get_write_concurrency to return
an error if XCom is stopping or has been stopped. Thus, we can avoid
parallel queries when we know that we will never have a proper response.

Change-Id: I8fe123f1ae2d9332be88179eb0d1a18ba04b6f4b