Bug#34876468 Missing ndb_schema uuid
On the NDB binlog thread start up steps, it checks if the `ndb_schema`
table definition in DD contains metadata indicating that it is
connecting to the same previous cluster. This check consists on
comparing the `schema_uuid` value in the `se_private_data` field with
the `query` column of the NDB table.

`ndb_restore` tool does not restore system table data, therefore the
`ndb_schema` on an initial data node start is empty. Since it is
important for the `ndb_schema` table to have such value for NDB binlog
startup, a new sanity check is performed against `ndb_schema`
upgrading the table, with a newly generated uuid, if necessary.

Change-Id: I811ddd7ca2c336e955d127ab6eed7c9f44a45f54