Bug #34670701 Too many ROLLUP rows with hypergraph. [1/4]
This commits adds a pool of temporary OrderElement array objects to
LogicalOrderings. This makes the code a bit more readable, as we no
longer need to pass these temporary objects down the call stack.  Also,
it means that we can allocate less memory, as these objects are now
reused more efficiently.

(This is a separate commit, to make the actual bugfix easier to read.)

Change-Id: I5ffa107eb7757feb12de619a7dcf7349ca7007ba