Bug#34927110: Failure in Item_ref::real_item, count_field_types
The problem is a missing reference count increment for an Item that
is added as part of subquery-to-derived transformation, which is
part of a predicate that is later eliminated.

Another issue was an Item_field object that was replaced in the
select list of a derived table, and subsequently had its reference count
decremented when checking for unused columns in derived tables.

All instances of adding and replacing expressions in select list when
transforming subqueries to use derived tables and joins have been
changed so that reference counts are maintained properly.

Change-Id: I1c6f528d82d55ea1f3fcd5ec902b71825cbe71d2