Bug#34977233 excessive memory copy in recv-buffer handling
While profiling it was seen that quite a bit of time is spent moving
memory.

That was caused by the assumption that the receive buffer always
contains the currently processed data at position 0.

In case of long results with many short rows, the receive buffer of 4k
may contain 512 rows, each a 8 bytes long. After every row that gets
forwarded to the send buffers and the receive buffer's "next frame at
position 0"-rule needs to be satisfied.

First 4088 (4096-8) bytes to the left, then 4080, 4072, ... On average
each move moves half the receive-buffer.

  512 moves * 2048 bytes = 1 Mbyte is moved.

Increasing the receive buffer to 64k would lead to:

  64kbyte / 8 byte/row = 8192 rows
  8192 moves * 64k/2   = 256 Mbytes

receive-buffer | moved bytes | moved-bytes / receive buffer
            4k |   1 * 1024k |                          256
           64k | 256 * 1024k |                         4096

The larger the receive buffer, the larger the amplification of the moved
bytes, which is opposite of what's intended.

Change
------

Instead of moving the bytes inside the recv-buffer to the front,
create a view over the recv-buffer and move the view to the back which
is a O(1) operation.

- renamed recv_plain_buffer() to recv_plain_view()
- replaced all uses of net::dynamic_buffer().consume(recv_plain_buf)
  (which moves bytes to the front) with .consume_plain()
  (which moves the view to the back)
- in case of unencrypted connections, let the recv_plain_buffer() refer
  to the raw network-buffers instead, avoiding another copy operation.

Change-Id: I131ca9895f6aa2668a9c3e7b447c9a7047ecd1f4