Bug #34946579 i_innodb.innodb-alter : Retry attempts for reading partial data failed
The Builder object is used to build one index.  In the destructor
of Builder object, wait for all the spawned parallel index build
threads to complete before closing the temporary files created.

During the Builder::State::FULL_SORT phase of building an index,
the Merge_cursor now does duplicate key checks.  This was not done
previously.  Now, since the data is already sorted and duplicates
are checked before reaching the build phase, during the build
phase it is not necessary to check the sort order of records.

Change-Id: I93c6c319412bbb22fbbdcc7a2ffd9e32fc87554b