Bug#34969838 heap-buffer-overflow in StrXfrmTest.ChineseUTF8MB4 gunit test
The loading/unloading of UCA character sets was never really written
to tolerate a sequence of init/uninit/init/uninit.

Rewrite handling of MY_UCA_INFO structs, so that it is simple to see
whether it owns allocated memory, or points to static data structures.

Re-introduce mem_malloc/mem_free in MY_CHARSET_LOADER, so that we can
free allocated memory. The previous use of once_alloc was really only
hiding memory leaks.

Change-Id: I02c792e612fe1bb3488ebe4161ed791fbe4aebc2