Bug#33657570 routertest_component_metadata_ttl unit test failed on daily trunk
Two issues reported in this bug:

1) Quarantine test failing in gcov build.
The expected log messages that indicate that the node is been quarantined
are not found in the log.

The problem is the race condition in the test between the connection attempt
that should cause the quarantine and the metadata refresh that is asynchronous.
When the refresh happens before the connection attempt the Router shuts down
the accepting port as it cannot connect to a single metadata server. In that
case there is no quarantine happening.

This patch changes the test so that there are 2 cluster nodes and the
refresh never causes the port shutdown as one of them is still accessible
after the other one was killed.

2) Metadata refresh cycle taking longer than the test assumes.

Given that the margins in the test are hundreds of milliseconds this test
cannot reliably check for the upper bound on PB2. There will always be some
rare runs when things get slower and it will fail sporadically despite increasing
the margins.

This patch gets rid of the upper bound checking, only lower bound is checked.

Change-Id: I2f08de9f50711782fa5cd98357433287745cadb9