Bug#34891365: MySQL 8.0.29+ is slower than MySQL 8.0.28 with queries using JOINS
Post-push fix.

Implicitly grouped queries on tables in the CSV engine failed under
ASan if the tables contained BLOBs.

The original bug fix made AggregateIterator store the contents of all
the record buffers when its child iterator was exhausted, so that they
could be restored if the AggregateIterator was reinitialized. This
interacted badly with the special handling of BLOB memory in the CSV
engine. The CSV engine has a separate MEM_ROOT for BLOB memory, and it
clears it for reuse before reading a new row. But it clears it before
it checks if it has reached EOF, so when EOF is reached, the
Field_blob objects will point to freed memory. Trying to copy that
memory into the AggregateIterator's buffer makes ASan fail.

Fixed by clearing the CSV engine's BLOB MEM_ROOT for reuse only when
there actually is a new row that can reuse the memory. This makes the
record buffer contain valid data after EOF is seen; either it has the
data from the last row, or it has empty data if the table was empty.
Then it's safe to store and restore the record buffer even after EOF.

Change-Id: I9b5051b0a16933c687677db6c4fe761d1eef7b96