Bug#33830659: Assertion `field->table == table'               failed|sql/sql_tmp_table.cc
When a query has ROLLUP, temp table aggregation is disabled.
However, in the presence of generated columns, after
substituting expressions having generated columns in a query,
a call is made to count_field_types(). This function sets
"allow_group_via_temp_table" to true but does not reset it
back which results in optimizer choosing temp table
aggregation leading to problems later.
We now do not set "allow_group_via_temp_table" unconditionally
in count_field_types(). If needed, it sets it to false.
However, this alone does not fix the problem. Since the group
by expression having a generated column is substituted with a
hidden field, it can no more be used to compare while
substituting ROLLUP expressions with temp table fields. So,
we now use the "rollup_item" stored as part of the structure
used for GROUP BY (during resolving of ROLLUP).
Additonally, since count_field_types() does not set
"allow_group_via_temp_table" unconditionally, we no more need
to reset it in the presence of ROLLUP in other places too. Also,
after the removal of slices for ROLLUP, call to
count_field_types() is not needed anymore. Earlier, rollup
handling would sometimes replace fields with constants which
meant optimizer had to re-count the field types. Since it's
not done anymore, call is now removed.

Change-Id: I9d01352519fed78e075aa66819ca7695387fddec