Bug#34992157 The AIO::s_log unused since WL#12527
The s_log slots array in AIO system was no longer needed after
refactoring of redo log performed in WL#12527.
So, was the thread created for it and some instrumentation.

This patch cleans up the logic of creating the AIO system,
and doxygen for it, so it is easier to understand.
The only observable change is elimination of the unnecessary io thread
which was handling (non-existent) redo log requests.

This patch also removes mentions of innobase_start_or_create_for_mysql
function which no longer exists, and fixes compilation errors for
UNIV_COMPILE_TEST_FUNCS flag.

Change-Id: I91cf835a4c78630a1f63b6cb2504ecf70d99ecfc