Bug #34284186 : ALTER INSTANCE RELOAD TLS can get stuck in spinlock
Description:
------------
ALTER INSTANCE RELOAD TLS makes use of MyRcuLock::write_wait_and_delete in order
to swap out the Ssl_acceptor_context_data object. The way it works is by first
updating the global pointer to point to the new object, and then waiting until
number of readers reaches 0 before freeing the old object. If there's a steady
amount of incoming ssl connections being made, then the number of readers may
not reach 0 in a reasonable amount of time. In the meantime, the thread will be
stuck in the spinlock.

Fix:
----
The fix is to introduce a wait after setting the variable to new value and
before disposing the old one.

Note: This patch is contributed by Sinisa Milivojevic.

Change-Id: I65218ef885accf9d7ee5b737b202d8dcefc35627