Bug#34964509 Assertion failure: btr0load.cc:912:blk == nullptr || blk->was_freed()
In the bulk load code, pages are allocated without knowledge of
buffer pool.  Once a page is allocated, there is a debug assert to
ensure that the paticular page is not there in the buffer pool.
This assert failed.  If the buffer pool can contain unallocated
page (perhaps with some old stale data), then this assert will not
work.  Hence we remove this assert.  Instead, just ensure that this
page is evicted from buffer pool.

Change-Id: I30a2fbf1ba2a66074d10e7672a97fd6351ff7131