Bug#34837464: Incorrect Left Outer Join Execution
Hash outer join sometimes incorrectly matched NULL with non-NULL
values. This could happen if the non-NULL value was a DECIMAL zero or
an empty string with a non-padding collation.

The problem was that the hash table built by HashJoinIterator included
rows that were known to match no rows in the probe table because the
join column was NULL and the join predicate was a regular equality (=)
and not a NULL-safe equality (<=>). It was stored with an empty join
key, which happens to be the same as the join key created for empty
string and for DECIMAL zero. If the probe table had rows with an empty
string or DECIMAL zero in the join column, the hash join would find a
false match in the hash table and return incorrect results.

This only happened for left or right outer joins. And it is true that
outer joins should return all rows in the outer table, even rows with
NULL in one of the join columns. But the hash table is always built
from the inner table of the outer join in our implementation, so we do
not need to have these extra rows in the hash table.

Fixed by always dropping rows from the hash table when they are known
to have no possible matches in the probe table.

Change-Id: I625d36e9acdf2b5d410505be915ef1c29b5b4703