BUG#34960922: Exit early from heatwave recovery if MySQL boot fails
Problem:
When MySQL server restarts, if rapid_reload_on_restart is turned on,
RAPID plugin tries to reload data that was previously loaded. To fetch
the list of tables, it needs to query information_schema.tables. For
this, the plugin waits for server to complete boot. It keeps polling
the server_operational_state in a loop to check if boot is completed
marked by SERVER_OPERATING, or if server is shutting down marked by
SERVER_SHUTTING_DOWN.

However, if MySQL server starts the boot process but encounters some
error on the way and initiate a shutdown instead, the value of
server_operational_state is not changed from the initial value of
SERVER_BOOTING. As a result, the RAPID plugin is not made aware of
the new state of the server. It keeps on polling, until after 300
failed attempts, it gives up.

This causes MySQL server shutdown process to unnecessarily wait for
the RAPID plugin to downgrade from ON to OFF state, which in turn
waits for data reload thread to complete, which waits for the MySQL
server to complete boot - thus forming a circular dependency.

Until the RAPID plugin exhausts it's 300 attempts to wait for MySQL
server to be operational, the system gets stuck doing practically
nothing. Eventually, the plugin gives up by throwing the following
error and let the plugin downgrade to unblock and thus MySQL server
to complete shutdown.

  m_getExternalGsOperation failed: Server is taking too long to boot,
  wait_for_server_bootup timed out.

Solution:
As a solution, we set server_operational_state to SERVER_SHUTTING_DOWN
when MySQL server is aborting due to error (in function unireg_abort).
This allows other modules like RAPID plugin to be aware that server
is shutting down and they can take appropriate action accordingly.
In this case, the plugin exits its data reload job early. This causes
the MySQL server to proceed with shutdown faster.

Change-Id: I74dbba6df147374ea5b461e861f92c847a40cd84