Bug#35010707 routertest_router_config_generator connects to fixed 3306 port
The problem is the Session that is supposed to be mocked by the
MySQLSessionReplayer via DIM. But the Router code does not always
create the MySQLSession using DIM which causes the real session be used
sometimes which is not the UTs intention.

This patch removes the MySQLSessionReplayers completely and also the
tests that are using it. These are very time consuming to maintain and
not very useful. They are mostly replaced by the mysql_mock_server-based
tests at this point.
This patch also adds some new tests to fill some of the coverage gaps
that resulted from removing the UTs mentioned above.

Change-Id: I927af9c24c3d2c8f796c06e75c0cdeaea3cf7fb0