Bug#34861841: Assertion : `false' in (anonymous namespace)::CostingReceiver::ProposeAccessPath
When executing a CREATE TABLE AS SELECT statement with the hypergraph
optimizer enabled, LEX::using_hypergraph_optimizer was not set until
the SELECT statement was executed. However, the flag needs to be set
before resolving the statement, as the resolver uses the flag to
decide whether a particular semijoin transformation should be
performed. The hypergraph optimizer does not support this
transformation, and assert failures were seen during optimization if
the resolver had performed the transformation.

Fixed by setting the using_hypergraph_optimizer flag before preparing
the CREATE TABLE AS SELECT statement.

Change-Id: I643fb8b797742651c8cdfa9b279e51e200c5a868