Bug#34856343: Assertion 'join->best_read < DBL_MAX'
This fix is for a regression from bug#34347116.

In this bugfix, it was erroneously changed which tables with a range
predicate that could be eliminated from join execution. The table in
the current bug report is the inner-most table of a nested outer join
query, where the inner-most outer join is wrapped in a derived table.

The logic around whether to allow range optimizer analysis and
subsequently use range optimizer analysis to eliminate tables from join
is both complex and imprecise, and this patch attempts to clean up the
processing.

It is first noticed that any table with a useful index can benefit
from range predicate analysis, regardless of it being part of an
inner join, outer join, semi-join or anti-join operation.

Second, the use of the "impossible_range" property from range predicate
analysis is clarified by allowing a const table to be removed when the
table is used in an inner join operation, when it is an outer table in
an outer join, semi-join or anti-join operation, or it is the sole inner
table of an outer join operation.

In all other cases, the "impossible_range" property is ignored because
the mark_const_table() function call is not able to update the data
structures consistently.

Elaborate test cases have been added for range optimizer cases and
const table elimination.

Change-Id: I104736579a87910a381636cf668d31646e3f711c