Bug#34899286: Assertion `!OrderItemsReferenceUnavailableTables( path, used_tables_before_repla
A query using a window function hit an assertion failure in the
hypergraph optimizer during finalization of the query plan. The
assertion complained that the SORT access path used for the window
referenced a column that was not available in the temporary table.

Simplified query:

SELECT ROW_NUMBER() OVER (PARTITION BY t1.x)
FROM t1, t2 WHERE t1.x = t2.pk
GROUP BY t1.pk;

The chosen plan was:

-> Window aggregate: row_number() OVER (PARTITION BY t1.x )
    -> Sort: t2.pk
        -> Stream results
            -> Group (no aggregates)
                -> Nested loop inner join
                    -> Index scan on t1 using PRIMARY
                    -> Filter: (t1.x = t2.pk)
                        -> Index scan on t2 using PRIMARY

The problem was that it picked "Sort: t2.pk" in order to make the
input to the "Window aggregate" step sorted on t1.x. Whether it gets
sorted on t1.x or t2.pk should not make any difference, since the join
condition t1.x = t2.pk ensures they are equal at this point. However,
since t2.pk is not mentioned in the SELECT list, GROUP BY, ORDER BY or
WINDOW clause, it does not get added to the temporary table used in
the "Stream results" step, so sorting on the t2.pk column cannot be
done after this step.

Fixed by marking the interesting orders that use columns that are not
available after materialization as "sort_ahead_only", and excluding
those orderings when creating SORT paths for windowing.

For completeness, also exclude those orderings in
ApplyDistinctAndOrder(), in case the fix for bug#34727088 did not
cover all similar issues for DISTINCT and ORDER BY.

Change-Id: Icf1011b39e21980234549b9dcb379e7be330f964