Bug#35013880 Wrong result for AVG() OVER(ROWS BETWEEN 1 FOLLOWING AND UNBOUNDED FOLLOWING)
A regression from the fix for Bug#34431996. In that fix, we discovered
that that the state variable Window::m_rowno_in_frame had been
interpreted in two ways: as the number of the row in the frame we
are currently accessing for evaluation with copy_funcs, and sometimes
also as the number of rows in a frame. The latter (now wrong)
interpretation stemmed from the original commit branch, and this issue
shows a remaining example of its use for AVG. Remove this usage in
favor of relying on Item_sum_avg::m_count. This should be safe, since
we do rely already on Item_sum_avg::m_frame_null_count; i.e. we do
know the total number of rows in the frame as well as the number of
NULLs from keeping track when adding or removing row contributions
from the intermediate sum.

I also checked any remaining uses of the state variable
m_rowno_in_frame in the evaluation functions and the all seem to use
the correct interpretation of it.

Test added for the two relevant code paths (val_real, val_decimal).

Change-Id: I23c093f8f0589ce5487c7c0bf8cedaaa16ec464d