Bug#34861841: Assertion : `false' in (anonymous namespace)::CostingReceiver::ProposeAccessPath
Post-push fix.

The original fix made prepared statements stop using the hypergraph
optimizer for CREATE TABLE AS SELECT statements. This caused problems
for HeatWave, which rejects statements if they are not using the
hypergraph optimizer, and rapid.ps_dict and rapid.ps_varlen fail as a
result of this.

It happened because the original fix set the
using_hypergraph_optimizer flag in the code path that prepared regular
statements, but not when preparing prepared statements. Fixed by
setting the using_hypergraph_optimizer flag also during preparation of
prepared statements.

Change-Id: I3fc70dc511d4323a4cdf57716b4e51fe96ecb492