Bug#35014853: Hypergraph no longer considers semi-to-inner when join condition is degenerate
Semijoins with degenerate join predicates did not consider
semijoin-to-inner transformation after bug#34670701.

This patch fixes that by making BuildInterestingOrders() set
ordering_idx_needed_for_semijoin_rewrite to the empty ordering for
such predicates.

Even after that fix, the hypergraph optimizer preferred to rewrite the
semijoin to an inner hash join instead of an inner nested loop join
when the tables were big. For the case where the inner table of the
semijoin (which is the outer table after rewriting to inner) is
deduplicated by using LIMIT 1. Then a nested loop join should
definitely be better, or at least no worse, than a hash join, as it
performs the same table scans but without having to build a hash
table.

This second problem was caused by ProposeHashJoin() and
ProposeNestedLoopJoin() using the cost and cardinality of the original
non-deduplicated "left_path". The cost would be calculated as if there
was a Cartesian product between the two tables. This patch updates the
cost calculations to use the deduplicated path to calculate the costs,
which makes the cost of the joins much lower, and the cost of the
nested loop join lower than the hash join.

Change-Id: I8212501469105732fa9841db17f6628fd41a7b96