Bug#33069747 ASSERTION `M_COUNT >= M_FRAME_NULL_COUNT && M_FRAME_NULL_COUNT > 0' FAILED
Missing call to split_sum func caused error when accessing rows from a
window frame of a window function call only present in the ORDER BY
list. This means that the argument of the window function was never
put in the select list nor never replaced with an Item_aggregate_ref
to it (in base_ref_items), cf. what happens if the window function is
part of an expression in the select list.

Change-Id: I287b701c48d0f96e57acba93b98f0312394a53c3