Bug#34370673: Failure in Item_subselect::print()
The query that causes the issue here is an anti-join operation
(originally a NOT IN subquery predicate), that contains another
IN subquery predicate as part of the WHERE clause of its subquery.
The contained IN subquery predicate is further part of an equality,
causing it to be decorrelated and added to sj_inner_exprs of the
anti-join nest.

During join simplification it is detected that the join condition of
the anti-join can be simplified to a FALSE operation, which also means
that the entire condition, including the IN subquery predicate,
is removed. However, the decorrelated predicate references are never
removed and some code traverses these structures during optimization
and thus causes a bad pointer traversal.

We actually try to eliminate these decorrelated references during
join simplification, but apparently the code is located at the
wrong place. The fix to the problem is to check the join condition
after simplification, and if it has been converted to a constant
value (either TRUE or FALSE), we know that any decorrelated
expressions can be eliminated.

In addition, some code in Query_block::simplify_joins() has been
adjusted to follow the current code standard more closely.

Change-Id: I515d2e4536923d678b89cbf883f2088e650a956b