Bug#34917498 ndb_import silently ignore hint failure for some big varchar primary key tables
Background:

ndb_import can use partitioning key from row for hinting which
transaction coordinator to use.

When it computes the hash for the key it have a 4092 byte buffer.

If the key do not have a character column the buffer will always be big
enough when calculating the hash.

Problem:

If a part of the key is a character column using utf8 the need for hash
buffer may be up to 24x the number of maximum characters in the column
depending on the collation in use.

If that is the case the hash computation fails but the calling code in
ndb_import does not check that and proceeds with an undefined hash value
causing an undefined hinting to occur.

That will not cause any functional problem, but it will be suboptimal
and also without notifying the user about it.

Another case is if the character column fits in the buffer to hash but
using more memory than the one byte per max characters, any following
column in a multi column key, may write beyond end of buffer since it is
assummed that key prepared for hashing will always fit in provided
buffer.

Fix:

ndb_import will use a buffer big enough also big character columns in
key for any collation.

And bounds checking are added in computeHash function to detect when
buffer is too small.

And check is added in ndb_import whether hash computation fails or not
and failure will be reported to user.

For testing a mode that uses smaller buffer is added
(--errins-type=bug34917498).

Change-Id: Ie4da13e47a2311c0a81e04f9b7bc83f2958e560c