BUG#35040290: Fix logic to allow DDL on tables with secondary engine
Problem:
Replication from HeatWave to MDS ReadReplica is broken when a DDL is run
on a table that is marked with SECONDARY_ENGINE=RAPID.
Since its beginning, DDLs on tables marked with SECONDARY_ENGINE were
blocked. After WL#15129, many DDLs were allowed on such tables. However,
it requires RAPID plugin to be installed. In Read-Replicas, RAPID plugin
is not installed because binary log is not enabled there. As a result,
if a table is marked as secondary_engine=rapid and an ALTER TABLE like
DDL is executed on it, the DDL succeeds on the source HeatWave, but
fails on the replicas with error:
  "DDLs on a table with a secondary engine defined are not allowed"

Solution:
In this commit, we are fixing the logic to allow/disallow DDLs on tables
with secondary engine. In case the secondary engine plugin itself is not
loaded (installed), then DDLs are allowed.
Currently, RAPID is the only secondary engine in production. When RAPID
plugin is not installed, then table marked with SECONDARY_ENGINE = RAPID
is no different than a regular InnoDB table. Even if the DDL does a
schema change that makes the table to be incompatible with RAPID, the
subsequent SECONDARY_LOAD command will fail with proper errors. And user
will be forced to take necessary action.

Note:
MTR tests are adjusted to adapt the new behavior.
Test main.truncate was expecting truncate table to fail because of
SECONDARY_ENGINE. Since this is no longer valid, the truncate table code
code is now adjusted to throw a simulated ER_SECONDARY_ENGINE error if a
certain dbug flag is found to be set by the test. The test is further
split into two parts as we need debug builds for simulating errors.

Test rapid.datalake_define is also adjusted accordingly.

Change-Id: I8aa239a91c84f2fe80a153bdab3812d913b7c1fa