WL#15582 ROUTER SET trace [3/4] query
ROUTER SET trace = 1 enables the trace for all following commands.

ROUTER SET trace = 1, change-user and reset-connection disable it.

The query attribute 'router.trace' enables/disables it for the current
statement.

The trace is stored in the diagnostics area and can be queries with SHOW
WARNINGS.

Multi-statement support is forbidden when connection-sharing is enabled.

Change-Id: Iaba81f57c8ea1ca501f213945b77197de01c4f22