Bug#34998721 Mysqld crash - Assertion `table_num_to_node_num[table_num] != -1' failed.
We hit an assert with the hypergraph optimizer, a table map number was
outside range (0 <= n < the number of tables of the query block),
after transformation of subqueries into derived tables: when the
original query has both an implicit grouping and a scalar subquery, we
move the original top level block into a derived table to do the
grouping there, cf. `Query_block::transform_grouped_to_derived' before
we make a derived table for the scalar subqueries, synthesizing a new
top level block which retains any window functions and order by
clauses.

In the test case, there is a top level window function with an ORDER
BY specification, referencing a top level field, t3.x, which is moved
into a derived table, by `transform_grouped_to_derived'. The ORDER BY
reference of the window is correctly updated to point to
`derived_1_4.Name_exp_1', but the reference is present as an argument
of a function call (`-'), originally as

        ORDER BY -t3.x         and after the transform as
        ORDER BY -derived_1_4.Name_exp_1

However, the Item_func_neg representing the monadic minus operator,
hadn't had its used_tables cache updated, which caused the range error
for the hypergraph optimizer.

The solution is to always update the used_tables cache when replacing
fields in expressions of the top level query block in
`transform_grouped_to_derived'.

Note: the window (an implicit one) isn't even used here, since it
occurs in the ORDER BY of an implicitly grouped query, so it is
eliminated. But the window definition is still attached to the query
block causing the hypergraph optimizer to consider its ORDER BY for
interesting orders and hence we see the error. We should probably
consider introducing reference counting for window definitions so they
can be removed when their use is gone.

Change-Id: Iaac93218b2bfbaba5c1adcc04d53436d68256373