Bug #33680100 maxRecordSize required to update in BACKUP_FRAGMENT_REQ for LCPs
Background:
When an ALTER TABLE adds columns to an existing table the maxRecordSize
used by LCPs to allocate buffer space for rows may change.
maxRecordSize is set in GET_TABINFOCONF and used later in BACKUP_FRAGMENT_REQ.

Problem:
If during the gap between the two signals an ALTER table change the
number of columns a wrong value of maxRecordSize may be used.

Fix:
Always update maxRecordSize on BACKUP_FRAGMENT_REQ (updated value of
maxRecordSize is read from TUP).

New API test TableAddAttrsUpdateMaxRecSzForLCP added.
Test uses the existing TableAddAttrs test to perform an ALTER TABLE and
also sets CFG_DB_LCP_INTERVAL to 0 in order to increase the chances
to have the ALTER TABLE, changing the number of columns of a table,
between GET_TABINFOCONF and BACKUP_FRAGMENT_REQ when one LCP is performed.

Change-Id: Icee7b84a8cf7313916df08cb0f1bb49f4d88a87a